+++ src/Exception/EntityRemovalBlockedException.php
@@ -6,7 +6,7 @@ namespace Becklyn\RadBundle\Exception;
  * Generic exception that marks, that the entity removal may be blocked by database (foreign keys, etc..) or
  * semantic constraints (e.g. "can't remove main category").
  */
-class EntityRemovalBlockedException extends \DomainException
+class EntityRemovalBlockedException extends RadException
 {
     /**
      * @var object[]
@@ -21,7 +21,7 @@ class EntityRemovalBlockedException extends \DomainException
      */
     public function __construct ($entities, string $message, \Throwable $previous = null)
     {
-        parent::__construct($message, 0, $previous);
+        parent::__construct($message, $previous);
         $this->entities = \is_array($entities) ? $entities : [$entities];
     }